+++ src/app/Models/Csvrow.php
